@@ -1,51 +1,41 @@
 name: flutter_html
 description: A Flutter widget rendering static HTML and CSS as Flutter widgets Forked from Sub6Resources
-version: 2.2.1
-homepage: https://github.com/hammadparveez/flutter_html
+version: 2.2.2
+homepage: https://github.com/coindcx-app/flutter_html
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: '>=2.2.0'
+  sdk: ^3.5.3
+  flutter: ">=1.17.0"
 
 dependencies:
   # Plugin for parsing html
-  html: '>=0.15.0 <1.0.0'
+  html: ^0.15.4
 
   # Plugins for parsing css
-  csslib: '>=0.17.0 <1.0.0'
+  csslib: ^1.0.0
 
   # Plugins for rendering the <table> tag.
-  flutter_layout_grid: '>=1.0.1 <2.0.0'
+  flutter_layout_grid: ^2.0.7
 
   # Plugins for rendering the <video> tag.
-  video_player: '>=2.1.1 <3.0.0'
-  chewie: '>=1.1.0 <2.0.0'
+  video_player: ^2.9.2
+  chewie: ^1.8.5
 
   # Plugin for rendering the <iframe> tag.
-#  webview_flutter: '>=2.0.4 <3.0.0'
-  webview_flutter: ^4.0.1
-
+  webview_flutter: ^4.9.0
 
   # Plugins for rendering the <audio> tag.
-  chewie_audio: '>=1.2.0 <2.0.0'
-
-
-
-  # Plugin for rendering MathML
-  flutter_math_fork: '>=0.6.3+1 <1.0.0'
-
+  chewie_audio: ^1.5.0
 
   # plugin for firstWhereOrNull extension on lists
-  collection: '>=1.15.0 <2.0.0'
+  collection: ^1.18.0
 
   # plugin to convert integers to numerals
-  numerus: '>=1.1.1 <2.0.0'
+  numerus: ^2.2.0
 
   flutter:
     sdk: flutter
-
-dependency_overrides:
-  flutter_svg: ^2.0.1
+  flutter_svg: ^2.0.10+1
 
 dev_dependencies:
   flutter_test:
