@@ -6,30 +6,24 @@ import 'package:csslib/parser.dart' as cssparser;
 import 'package:csslib/visitor.dart' as css;
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter_html/flutter_html.dart';
-import 'package:flutter_html/image_render.dart';
-import 'package:flutter_html/src/anchor.dart';
 import 'package:flutter_html/src/css_parser.dart';
 import 'package:flutter_html/src/html_elements.dart';
-import 'package:flutter_html/src/layout_element.dart';
-import 'package:flutter_html/src/navigation_delegate.dart';
 import 'package:flutter_html/src/utils.dart';
-import 'package:flutter_html/style.dart';
 import 'package:html/dom.dart' as dom;
 import 'package:html/parser.dart' as htmlparser;
 import 'package:numerus/numerus.dart';
 
 typedef OnTap = void Function(
-    String? url,
-    RenderContext context,
-    Map<String, String> attributes,
-    dom.Element? element,
+  String? url,
+  RenderContext context,
+  Map<String, String> attributes,
+  dom.Element? element,
 );
 typedef OnMathError = Widget Function(
-    String parsedTex,
-    String exception,
-    String exceptionWithType,
+  String parsedTex,
+  String exception,
+  String exceptionWithType,
 );
 typedef OnCssParseError = String? Function(
   String css,
@@ -81,16 +75,15 @@ class HtmlParser extends StatelessWidget {
     required this.navigationDelegateForIframe,
     this.selectionControls,
     this.scrollPhysics,
-  })  : this._onAnchorTap = onAnchorTap != null
-          ? onAnchorTap
-          : key != null
-              ? _handleAnchorTap(key, onLinkTap)
-              : null,
+  })  : _onAnchorTap = onAnchorTap ??
+            (key != null ? _handleAnchorTap(key, onLinkTap) : null),
         super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    Map<String, Map<String, List<css.Expression>>> declarations = _getExternalCssDeclarations(htmlData.getElementsByTagName("style"), onCssParseError);
+    Map<String, Map<String, List<css.Expression>>> declarations =
+        _getExternalCssDeclarations(
+            htmlData.getElementsByTagName("style"), onCssParseError);
     StyledElement lexedTree = lexDomTree(
       htmlData,
       customRender.keys.toList(),
@@ -102,8 +95,10 @@ class HtmlParser extends StatelessWidget {
     if (declarations.isNotEmpty) {
       externalCssStyledTree = _applyExternalCss(declarations, lexedTree);
     }
-    StyledElement inlineStyledTree = _applyInlineStyles(externalCssStyledTree ?? lexedTree, onCssParseError);
-    StyledElement customStyledTree = _applyCustomStyles(style, inlineStyledTree);
+    StyledElement inlineStyledTree =
+        _applyInlineStyles(externalCssStyledTree ?? lexedTree, onCssParseError);
+    StyledElement customStyledTree =
+        _applyCustomStyles(style, inlineStyledTree);
     StyledElement cascadedStyledTree = _cascadeStyles(style, customStyledTree);
     StyledElement cleanedTree = cleanTree(cascadedStyledTree);
     InlineSpan parsedTree = parseTree(
@@ -124,7 +119,7 @@ class HtmlParser extends StatelessWidget {
       return StyledText.selectable(
         textSpan: parsedTree as TextSpan,
         style: cleanedTree.style,
-        textScaleFactor: MediaQuery.of(context).textScaleFactor,
+        textScaler: MediaQuery.of(context).textScaler,
         renderContext: RenderContext(
           buildContext: context,
           parser: this,
@@ -138,7 +133,7 @@ class HtmlParser extends StatelessWidget {
     return StyledText(
       textSpan: parsedTree,
       style: cleanedTree.style,
-      textScaleFactor: MediaQuery.of(context).textScaleFactor,
+      textScaler: MediaQuery.of(context).textScaler,
       renderContext: RenderContext(
         buildContext: context,
         parser: this,
@@ -170,17 +165,17 @@ class HtmlParser extends StatelessWidget {
       name: "[Tree Root]",
       children: <StyledElement>[],
       node: html.documentElement,
-      style: Style.fromTextStyle(Theme.of(context).textTheme.bodyText2!),
+      style: Style.fromTextStyle(Theme.of(context).textTheme.bodyMedium!),
     );
 
-    html.nodes.forEach((node) {
+    for (var node in html.nodes) {
       tree.children.add(_recursiveLexer(
         node,
         customRenderTags,
         tagsList,
         navigationDelegateForIframe,
       ));
-    });
+    }
 
     return tree;
   }
@@ -197,14 +192,14 @@ class HtmlParser extends StatelessWidget {
   ) {
     List<StyledElement> children = <StyledElement>[];
 
-    node.nodes.forEach((childNode) {
+    for (var childNode in node.nodes) {
       children.add(_recursiveLexer(
         childNode,
         customRenderTags,
         tagsList,
         navigationDelegateForIframe,
       ));
-    });
+    }
 
     //TODO(Sub6Resources): There's probably a more efficient way to look this up.
     if (node is dom.Element) {
@@ -216,7 +211,8 @@ class HtmlParser extends StatelessWidget {
       } else if (INTERACTABLE_ELEMENTS.contains(node.localName)) {
         return parseInteractableElement(node, children);
       } else if (REPLACED_ELEMENTS.contains(node.localName)) {
-        return parseReplacedElement(node, children, navigationDelegateForIframe);
+        return parseReplacedElement(
+            node, children, navigationDelegateForIframe);
       } else if (LAYOUT_ELEMENTS.contains(node.localName)) {
         return parseLayoutElement(node, children);
       } else if (TABLE_CELL_ELEMENTS.contains(node.localName)) {
@@ -229,13 +225,16 @@ class HtmlParser extends StatelessWidget {
         return EmptyContentElement();
       }
     } else if (node is dom.Text) {
-      return TextContentElement(text: node.text, style: Style(), element: node.parent, node: node);
+      return TextContentElement(
+          text: node.text, style: Style(), element: node.parent, node: node);
     } else {
       return EmptyContentElement();
     }
   }
 
-  static Map<String, Map<String, List<css.Expression>>> _getExternalCssDeclarations(List<dom.Element> styles, OnCssParseError? errorHandler) {
+  static Map<String, Map<String, List<css.Expression>>>
+      _getExternalCssDeclarations(
+          List<dom.Element> styles, OnCssParseError? errorHandler) {
     String fullCss = "";
     for (final e in styles) {
       fullCss = fullCss + e.innerHtml;
@@ -248,7 +247,9 @@ class HtmlParser extends StatelessWidget {
     }
   }
 
-  static StyledElement _applyExternalCss(Map<String, Map<String, List<css.Expression>>> declarations, StyledElement tree) {
+  static StyledElement _applyExternalCss(
+      Map<String, Map<String, List<css.Expression>>> declarations,
+      StyledElement tree) {
     declarations.forEach((key, style) {
       try {
         if (tree.matchesSelector(key)) {
@@ -257,12 +258,15 @@ class HtmlParser extends StatelessWidget {
       } catch (_) {}
     });
 
-    tree.children.forEach((e) => _applyExternalCss(declarations, e));
+    for (var e in tree.children) {
+      _applyExternalCss(declarations, e);
+    }
 
     return tree;
   }
 
-  static StyledElement _applyInlineStyles(StyledElement tree, OnCssParseError? errorHandler) {
+  static StyledElement _applyInlineStyles(
+      StyledElement tree, OnCssParseError? errorHandler) {
     if (tree.attributes.containsKey("style")) {
       final newStyle = inlineCssToStyle(tree.attributes['style'], errorHandler);
       if (newStyle != null) {
@@ -270,13 +274,16 @@ class HtmlParser extends StatelessWidget {
       }
     }
 
-    tree.children.forEach((e) => _applyInlineStyles(e, errorHandler));
+    for (var e in tree.children) {
+      _applyInlineStyles(e, errorHandler);
+    }
     return tree;
   }
 
   /// [applyCustomStyles] applies the [Style] objects passed into the [Html]
   /// widget onto the [StyledElement] tree, no cascading of styles is done at this point.
-  static StyledElement _applyCustomStyles(Map<String, Style> style, StyledElement tree) {
+  static StyledElement _applyCustomStyles(
+      Map<String, Style> style, StyledElement tree) {
     style.forEach((key, style) {
       try {
         if (tree.matchesSelector(key)) {
@@ -284,18 +291,21 @@ class HtmlParser extends StatelessWidget {
         }
       } catch (_) {}
     });
-    tree.children.forEach((e) => _applyCustomStyles(style, e));
+    for (var e in tree.children) {
+      _applyCustomStyles(style, e);
+    }
 
     return tree;
   }
 
   /// [_cascadeStyles] cascades all of the inherited styles down the tree, applying them to each
   /// child that doesn't specify a different style.
-  static StyledElement _cascadeStyles(Map<String, Style> style, StyledElement tree) {
-    tree.children.forEach((child) {
+  static StyledElement _cascadeStyles(
+      Map<String, Style> style, StyledElement tree) {
+    for (var child in tree.children) {
       child.style = tree.style.copyOnlyInherited(child.style);
       _cascadeStyles(style, child);
-    });
+    }
 
     return tree;
   }
@@ -336,7 +346,8 @@ class HtmlParser extends StatelessWidget {
           newContext: newContext,
           style: tree.style,
           shrinkWrap: context.parser.shrinkWrap,
-          children: tree.children.map((tree) => parseTree(newContext, tree)).toList(),
+          children:
+              tree.children.map((tree) => parseTree(newContext, tree)).toList(),
         ),
       );
       if (render != null) {
@@ -363,17 +374,17 @@ class HtmlParser extends StatelessWidget {
           style: newContext.style.generateTextStyle(),
           children: tree.children
               .expandIndexed((i, childTree) => [
-            if (childTree.style.display == Display.BLOCK &&
-                i > 0 &&
-                tree.children[i - 1] is ReplacedElement)
-              TextSpan(text: "\n"),
-            parseTree(newContext, childTree),
-            if (i != tree.children.length - 1 &&
-                childTree.style.display == Display.BLOCK &&
-                childTree.element?.localName != "html" &&
-                childTree.element?.localName != "body")
-              TextSpan(text: "\n"),
-          ])
+                    if (childTree.style.display == Display.BLOCK &&
+                        i > 0 &&
+                        tree.children[i - 1] is ReplacedElement)
+                      TextSpan(text: "\n"),
+                    parseTree(newContext, childTree),
+                    if (i != tree.children.length - 1 &&
+                        childTree.style.display == Display.BLOCK &&
+                        childTree.element?.localName != "html" &&
+                        childTree.element?.localName != "body")
+                      TextSpan(text: "\n"),
+                  ])
               .toList(),
         );
       }
@@ -403,10 +414,13 @@ class HtmlParser extends StatelessWidget {
       );
     } else if (tree.style.display == Display.LIST_ITEM) {
       List<InlineSpan> getChildren(StyledElement tree) {
-        List<InlineSpan> children = tree.children.map((tree) => parseTree(newContext, tree)).toList();
+        List<InlineSpan> children =
+            tree.children.map((tree) => parseTree(newContext, tree)).toList();
         if (tree.style.listStylePosition == ListStylePosition.INSIDE) {
           final tabSpan = WidgetSpan(
-            child: Text("\t", textAlign: TextAlign.right, style: TextStyle(fontWeight: FontWeight.w400)),
+            child: Text("\t",
+                textAlign: TextAlign.right,
+                style: TextStyle(fontWeight: FontWeight.w400)),
           );
           children.insert(0, tabSpan);
         }
@@ -424,29 +438,54 @@ class HtmlParser extends StatelessWidget {
             mainAxisSize: MainAxisSize.min,
             textDirection: tree.style.direction,
             children: [
-              tree.style.listStylePosition == ListStylePosition.OUTSIDE ?
-              Padding(
-                padding: tree.style.padding?.nonNegative ?? EdgeInsets.only(left: tree.style.direction != TextDirection.rtl ? 10.0 : 0.0, right: tree.style.direction == TextDirection.rtl ? 10.0 : 0.0),
-                child: newContext.style.markerContent
-              ) : Container(height: 0, width: 0),
-              Text("\t", textAlign: TextAlign.right, style: TextStyle(fontWeight: FontWeight.w400)),
+              tree.style.listStylePosition == ListStylePosition.OUTSIDE
+                  ? Padding(
+                      padding: tree.style.padding?.nonNegative ??
+                          EdgeInsets.only(
+                              left: tree.style.direction != TextDirection.rtl
+                                  ? 10.0
+                                  : 0.0,
+                              right: tree.style.direction == TextDirection.rtl
+                                  ? 10.0
+                                  : 0.0),
+                      child: newContext.style.markerContent)
+                  : SizedBox(height: 0, width: 0),
+              Text("\t",
+                  textAlign: TextAlign.right,
+                  style: TextStyle(fontWeight: FontWeight.w400)),
               Expanded(
                   child: Padding(
-                      padding: tree.style.listStylePosition == ListStylePosition.INSIDE ?
-                        EdgeInsets.only(left: tree.style.direction != TextDirection.rtl ? 10.0 : 0.0, right: tree.style.direction == TextDirection.rtl ? 10.0 : 0.0) : EdgeInsets.zero,
+                      padding: tree.style.listStylePosition ==
+                              ListStylePosition.INSIDE
+                          ? EdgeInsets.only(
+                              left: tree.style.direction != TextDirection.rtl
+                                  ? 10.0
+                                  : 0.0,
+                              right: tree.style.direction == TextDirection.rtl
+                                  ? 10.0
+                                  : 0.0)
+                          : EdgeInsets.zero,
                       child: StyledText(
                         textSpan: TextSpan(
-                          children: getChildren(tree)..insertAll(0, tree.style.listStylePosition == ListStylePosition.INSIDE ?
-                            [
-                              WidgetSpan(alignment: PlaceholderAlignment.middle, child: newContext.style.markerContent ?? Container(height: 0, width: 0))
-                            ] : []),
+                          children: getChildren(tree)
+                            ..insertAll(
+                                0,
+                                tree.style.listStylePosition ==
+                                        ListStylePosition.INSIDE
+                                    ? [
+                                        WidgetSpan(
+                                            alignment:
+                                                PlaceholderAlignment.middle,
+                                            child: newContext
+                                                    .style.markerContent ??
+                                                SizedBox(height: 0, width: 0))
+                                      ]
+                                    : []),
                           style: newContext.style.generateTextStyle(),
                         ),
                         style: newContext.style,
                         renderContext: context,
-                      )
-                  )
-              )
+                      )))
             ],
           ),
         ),
@@ -476,8 +515,10 @@ class HtmlParser extends StatelessWidget {
                     : childStyle.merge(childSpan.style)),
             semanticsLabel: childSpan.semanticsLabel,
             recognizer: TapGestureRecognizer()
-              ..onTap =
-                  _onAnchorTap != null ? () => _onAnchorTap!(tree.href, context, tree.attributes, tree.element) : null,
+              ..onTap = _onAnchorTap != null
+                  ? () => _onAnchorTap(
+                      tree.href, context, tree.attributes, tree.element)
+                  : null,
           );
         } else {
           return WidgetSpan(
@@ -486,12 +527,14 @@ class HtmlParser extends StatelessWidget {
               cursor: SystemMouseCursors.click,
               child: MultipleTapGestureDetector(
                 onTap: _onAnchorTap != null
-                ? () => _onAnchorTap!(tree.href, context, tree.attributes, tree.element)
+                    ? () => _onAnchorTap(
+                        tree.href, context, tree.attributes, tree.element)
                     : null,
                 child: GestureDetector(
                   key: AnchorKey.of(key, tree),
                   onTap: _onAnchorTap != null
-                  ? () => _onAnchorTap!(tree.href, context, tree.attributes, tree.element)
+                      ? () => _onAnchorTap(
+                          tree.href, context, tree.attributes, tree.element)
                       : null,
                   child: (childSpan as WidgetSpan).child,
                 ),
@@ -504,11 +547,11 @@ class HtmlParser extends StatelessWidget {
       return TextSpan(
         mouseCursor: SystemMouseCursors.click,
         children: tree.children
-                .map((tree) => parseTree(newContext, tree))
-                .map((childSpan) {
+            .map((tree) => parseTree(newContext, tree))
+            .map((childSpan) {
           return addTaps(childSpan,
-            newContext.style.generateTextStyle().merge(childSpan.style));
-          }).toList(),
+              newContext.style.generateTextStyle().merge(childSpan.style));
+        }).toList(),
       );
     } else if (tree is LayoutElement) {
       return WidgetSpan(
@@ -535,7 +578,9 @@ class HtmlParser extends StatelessWidget {
           child: StyledText(
             textSpan: TextSpan(
               style: newContext.style.generateTextStyle(),
-              children: tree.children.map((tree) => parseTree(newContext, tree)).toList(),
+              children: tree.children
+                  .map((tree) => parseTree(newContext, tree))
+                  .toList(),
             ),
             style: newContext.style,
             renderContext: newContext,
@@ -559,10 +604,13 @@ class HtmlParser extends StatelessWidget {
     }
   }
 
-  static OnTap _handleAnchorTap(Key key, OnTap? onLinkTap) =>
-          (String? url, RenderContext context, Map<String, String> attributes, dom.Element? element) {
+  static OnTap _handleAnchorTap(Key key, OnTap? onLinkTap) => (String? url,
+          RenderContext context,
+          Map<String, String> attributes,
+          dom.Element? element) {
         if (url?.startsWith("#") == true) {
-          final anchorContext = AnchorKey.forId(key, url!.substring(1))?.currentContext;
+          final anchorContext =
+              AnchorKey.forId(key, url!.substring(1))?.currentContext;
           if (anchorContext != null) {
             Scrollable.ensureVisible(anchorContext);
           }
@@ -606,24 +654,33 @@ class HtmlParser extends StatelessWidget {
       /// initialize indices to negative numbers to make conditionals a little easier
       int textIndex = -1;
       int elementIndex = -1;
+
       /// initialize parent after to a whitespace to account for elements that are
       /// the last child in the list of elements
       String parentAfterText = " ";
+
       /// find the index of the text in the current tree
       if ((tree.element?.nodes.length ?? 0) >= 1) {
-        textIndex = tree.element?.nodes.indexWhere((element) => element == tree.node) ?? -1;
+        textIndex =
+            tree.element?.nodes.indexWhere((element) => element == tree.node) ??
+                -1;
       }
+
       /// get the parent nodes
       dom.NodeList? parentNodes = tree.element?.parent?.nodes;
+
       /// find the index of the tree itself in the parent nodes
       if ((parentNodes?.length ?? 0) >= 1) {
-        elementIndex = parentNodes?.indexWhere((element) => element == tree.element) ?? -1;
+        elementIndex =
+            parentNodes?.indexWhere((element) => element == tree.element) ?? -1;
       }
+
       /// if the tree is any node except the last node in the node list and the
       /// next node in the node list is a text node, then get its text. Otherwise
       /// the next node will be a [dom.Element], so keep unwrapping that until
       /// we get the underlying text node, and finally get its text.
-      if (elementIndex < (parentNodes?.length ?? 1) - 1 && parentNodes?[elementIndex + 1] is dom.Text) {
+      if (elementIndex < (parentNodes?.length ?? 1) - 1 &&
+          parentNodes?[elementIndex + 1] is dom.Text) {
         parentAfterText = parentNodes?[elementIndex + 1].text ?? " ";
       } else if (elementIndex < (parentNodes?.length ?? 1) - 1) {
         var parentAfter = parentNodes?[elementIndex + 1];
@@ -636,6 +693,7 @@ class HtmlParser extends StatelessWidget {
         }
         parentAfterText = parentAfter?.text ?? " ";
       }
+
       /// If the text is the first element in the current tree node list, it
       /// starts with a whitespace, it isn't a line break, either the
       /// whitespace is unnecessary or it is a block element, and either it is
@@ -645,38 +703,39 @@ class HtmlParser extends StatelessWidget {
       /// We should also delete the whitespace at any point in the node list
       /// if the previous element is a <br> because that tag makes the element
       /// act like a block element.
-      if (textIndex < 1
-          && tree.text!.startsWith(' ')
-          && tree.element?.localName != "br"
-          && (!keepLeadingSpace.data
-              || BLOCK_ELEMENTS.contains(tree.element?.localName ?? ""))
-          && (elementIndex < 1
-              || (elementIndex >= 1
-                  && parentNodes?[elementIndex - 1] is dom.Text
-                  && parentNodes![elementIndex - 1].text!.endsWith(" ")))
-      ) {
+      if (textIndex < 1 &&
+          tree.text!.startsWith(' ') &&
+          tree.element?.localName != "br" &&
+          (!keepLeadingSpace.data ||
+              BLOCK_ELEMENTS.contains(tree.element?.localName ?? "")) &&
+          (elementIndex < 1 ||
+              (elementIndex >= 1 &&
+                  parentNodes?[elementIndex - 1] is dom.Text &&
+                  parentNodes![elementIndex - 1].text!.endsWith(" ")))) {
         tree.text = tree.text!.replaceFirst(' ', '');
-      } else if (textIndex >= 1
-          && tree.text!.startsWith(' ')
-          && tree.element?.nodes[textIndex - 1] is dom.Element
-          && (tree.element?.nodes[textIndex - 1] as dom.Element).localName == "br"
-      ) {
+      } else if (textIndex >= 1 &&
+          tree.text!.startsWith(' ') &&
+          tree.element?.nodes[textIndex - 1] is dom.Element &&
+          (tree.element?.nodes[textIndex - 1] as dom.Element).localName ==
+              "br") {
         tree.text = tree.text!.replaceFirst(' ', '');
       }
+
       /// If the text is the last element in the current tree node list, it isn't
       /// a line break, and the next text node starts with a whitespace,
       /// update the [Context] to signify to that next text node whether it should
       /// keep its whitespace. This is based on whether the current text ends with a
       /// whitespace.
-      if (textIndex == (tree.element?.nodes.length ?? 1) - 1
-          && tree.element?.localName != "br"
-          && parentAfterText.startsWith(' ')
-      ) {
+      if (textIndex == (tree.element?.nodes.length ?? 1) - 1 &&
+          tree.element?.localName != "br" &&
+          parentAfterText.startsWith(' ')) {
         keepLeadingSpace.data = !tree.text!.endsWith(' ');
       }
     }
 
-    tree.children.forEach((e) => _processInlineWhitespaceRecursive(e, keepLeadingSpace));
+    for (var e in tree.children) {
+      _processInlineWhitespaceRecursive(e, keepLeadingSpace);
+    }
 
     return tree;
   }
@@ -690,8 +749,8 @@ class HtmlParser extends StatelessWidget {
   /// (4) Replace any instances of two or more spaces with a single space.
   static String _removeUnnecessaryWhitespace(String text) {
     return text
-        .replaceAll(RegExp("\ *(?=\n)"), "\n")
-        .replaceAll(RegExp("(?:\n)\ *"), "\n")
+        .replaceAll(RegExp(" *(?=\n)"), "\n")
+        .replaceAll(RegExp("(?:\n) *"), "\n")
         .replaceAll("\n", " ")
         .replaceAll("\t", " ")
         .replaceAll(RegExp(" {2,}"), " ");
@@ -710,17 +769,20 @@ class HtmlParser extends StatelessWidget {
   /// bullet all list items according to the [ListStyleType] they have been given.
   static StyledElement _processListCharactersRecursive(
       StyledElement tree, ListQueue<Context> olStack) {
-    if (tree.style.listStylePosition == null) {
-      tree.style.listStylePosition = ListStylePosition.OUTSIDE;
-    }
-    if (tree.name == 'ol' && tree.style.listStyleType != null && tree.style.listStyleType!.type == "marker") {
+    tree.style.listStylePosition ??= ListStylePosition.OUTSIDE;
+    if (tree.name == 'ol' &&
+        tree.style.listStyleType != null &&
+        tree.style.listStyleType!.type == "marker") {
       switch (tree.style.listStyleType!) {
         case ListStyleType.LOWER_LATIN:
         case ListStyleType.LOWER_ALPHA:
         case ListStyleType.UPPER_LATIN:
         case ListStyleType.UPPER_ALPHA:
           olStack.add(Context<String>('a'));
-          if ((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start']!) : null) != null) {
+          if ((tree.attributes['start'] != null
+                  ? int.tryParse(tree.attributes['start']!)
+                  : null) !=
+              null) {
             var start = int.tryParse(tree.attributes['start']!) ?? 1;
             var x = 1;
             while (x < start) {
@@ -730,14 +792,22 @@ class HtmlParser extends StatelessWidget {
           }
           break;
         default:
-          olStack.add(Context<int>((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start'] ?? "") ?? 1 : 1) - 1));
+          olStack.add(Context<int>((tree.attributes['start'] != null
+                  ? int.tryParse(tree.attributes['start'] ?? "") ?? 1
+                  : 1) -
+              1));
           break;
       }
-    } else if (tree.style.display == Display.LIST_ITEM && tree.style.listStyleType != null && tree.style.listStyleType!.type == "widget") {
+    } else if (tree.style.display == Display.LIST_ITEM &&
+        tree.style.listStyleType != null &&
+        tree.style.listStyleType!.type == "widget") {
       tree.style.markerContent = tree.style.listStyleType!.widget!;
-    } else if (tree.style.display == Display.LIST_ITEM && tree.style.listStyleType != null && tree.style.listStyleType!.type == "image") {
+    } else if (tree.style.display == Display.LIST_ITEM &&
+        tree.style.listStyleType != null &&
+        tree.style.listStyleType!.type == "image") {
       tree.style.markerContent = Image.network(tree.style.listStyleType!.text);
-    } else if (tree.style.display == Display.LIST_ITEM && tree.style.listStyleType != null) {
+    } else if (tree.style.display == Display.LIST_ITEM &&
+        tree.style.listStyleType != null) {
       String marker = "";
       switch (tree.style.listStyleType!) {
         case ListStyleType.NONE:
@@ -753,7 +823,10 @@ class HtmlParser extends StatelessWidget {
           break;
         case ListStyleType.DECIMAL:
           if (olStack.isEmpty) {
-            olStack.add(Context<int>((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start'] ?? "") ?? 1 : 1) - 1));
+            olStack.add(Context<int>((tree.attributes['start'] != null
+                    ? int.tryParse(tree.attributes['start'] ?? "") ?? 1
+                    : 1) -
+                1));
           }
           olStack.last.data += 1;
           marker = '${olStack.last.data}.';
@@ -762,7 +835,10 @@ class HtmlParser extends StatelessWidget {
         case ListStyleType.LOWER_ALPHA:
           if (olStack.isEmpty) {
             olStack.add(Context<String>('a'));
-            if ((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start']!) : null) != null) {
+            if ((tree.attributes['start'] != null
+                    ? int.tryParse(tree.attributes['start']!)
+                    : null) !=
+                null) {
               var start = int.tryParse(tree.attributes['start']!) ?? 1;
               var x = 1;
               while (x < start) {
@@ -771,14 +847,17 @@ class HtmlParser extends StatelessWidget {
               }
             }
           }
-          marker = olStack.last.data.toString() + ".";
+          marker = "${olStack.last.data}.";
           olStack.last.data = olStack.last.data.toString().nextLetter();
           break;
         case ListStyleType.UPPER_LATIN:
         case ListStyleType.UPPER_ALPHA:
           if (olStack.isEmpty) {
             olStack.add(Context<String>('a'));
-            if ((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start']!) : null) != null) {
+            if ((tree.attributes['start'] != null
+                    ? int.tryParse(tree.attributes['start']!)
+                    : null) !=
+                null) {
               var start = int.tryParse(tree.attributes['start']!) ?? 1;
               var x = 1;
               while (x < start) {
@@ -787,39 +866,48 @@ class HtmlParser extends StatelessWidget {
               }
             }
           }
-          marker = olStack.last.data.toString().toUpperCase() + ".";
+          marker = "${olStack.last.data.toString().toUpperCase()}.";
           olStack.last.data = olStack.last.data.toString().nextLetter();
           break;
         case ListStyleType.LOWER_ROMAN:
           if (olStack.isEmpty) {
-            olStack.add(Context<int>((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start'] ?? "") ?? 1 : 1) - 1));
+            olStack.add(Context<int>((tree.attributes['start'] != null
+                    ? int.tryParse(tree.attributes['start'] ?? "") ?? 1
+                    : 1) -
+                1));
           }
           olStack.last.data += 1;
           if (olStack.last.data <= 0) {
             marker = '${olStack.last.data}.';
           } else {
-            marker = (olStack.last.data as int).toRomanNumeralString()!.toLowerCase() + ".";
+            marker =
+                "${(olStack.last.data as int).toRomanNumeralString()!.toLowerCase()}.";
           }
           break;
         case ListStyleType.UPPER_ROMAN:
           if (olStack.isEmpty) {
-            olStack.add(Context<int>((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start'] ?? "") ?? 1 : 1) - 1));
+            olStack.add(Context<int>((tree.attributes['start'] != null
+                    ? int.tryParse(tree.attributes['start'] ?? "") ?? 1
+                    : 1) -
+                1));
           }
           olStack.last.data += 1;
           if (olStack.last.data <= 0) {
             marker = '${olStack.last.data}.';
           } else {
-            marker = (olStack.last.data as int).toRomanNumeralString()! + ".";
+            marker = "${(olStack.last.data as int).toRomanNumeralString()!}.";
           }
           break;
       }
       tree.style.markerContent = Text(
-          marker,
-          textAlign: TextAlign.right,
+        marker,
+        textAlign: TextAlign.right,
       );
     }
 
-    tree.children.forEach((e) => _processListCharactersRecursive(e, olStack));
+    for (var e in tree.children) {
+      _processListCharactersRecursive(e, olStack);
+    }
 
     if (tree.name == 'ol') {
       olStack.removeLast();
@@ -834,11 +922,17 @@ class HtmlParser extends StatelessWidget {
   static StyledElement _processBeforesAndAfters(StyledElement tree) {
     if (tree.style.before != null) {
       tree.children.insert(
-          0, TextContentElement(text: tree.style.before, style: tree.style.copyWith(beforeAfterNull: true, display: Display.INLINE)));
+          0,
+          TextContentElement(
+              text: tree.style.before,
+              style: tree.style
+                  .copyWith(beforeAfterNull: true, display: Display.INLINE)));
     }
     if (tree.style.after != null) {
-      tree.children
-          .add(TextContentElement(text: tree.style.after, style: tree.style.copyWith(beforeAfterNull: true, display: Display.INLINE)));
+      tree.children.add(TextContentElement(
+          text: tree.style.after,
+          style: tree.style
+              .copyWith(beforeAfterNull: true, display: Display.INLINE)));
     }
 
     tree.children.forEach(_processBeforesAndAfters);
@@ -958,16 +1052,16 @@ class HtmlParser extends StatelessWidget {
   static StyledElement _removeEmptyElements(StyledElement tree) {
     List<StyledElement> toRemove = <StyledElement>[];
     bool lastChildBlock = true;
-    tree.children.forEach((child) {
+    for (var child in tree.children) {
       if (child is EmptyContentElement || child is EmptyLayoutElement) {
         toRemove.add(child);
-      } else if (child is TextContentElement
-          && (tree.name == "body" || tree.name == "ul")
-          && child.text!.replaceAll(' ', '').isEmpty) {
+      } else if (child is TextContentElement &&
+          (tree.name == "body" || tree.name == "ul") &&
+          child.text!.replaceAll(' ', '').isEmpty) {
         toRemove.add(child);
-      } else if (child is TextContentElement
-          && child.text!.isEmpty
-          && child.style.whiteSpace != WhiteSpace.PRE) {
+      } else if (child is TextContentElement &&
+          child.text!.isEmpty &&
+          child.style.whiteSpace != WhiteSpace.PRE) {
         toRemove.add(child);
       } else if (child is TextContentElement &&
           child.style.whiteSpace != WhiteSpace.PRE &&
@@ -985,7 +1079,7 @@ class HtmlParser extends StatelessWidget {
       lastChildBlock = (child.style.display == Display.BLOCK ||
           child.style.display == Display.LIST_ITEM ||
           (child is TextContentElement && child.text == '\n'));
-    });
+    }
     tree.children.removeWhere((element) => toRemove.contains(element));
 
     return tree;
@@ -996,14 +1090,14 @@ class HtmlParser extends StatelessWidget {
   static StyledElement _processFontSize(StyledElement tree) {
     double? parentFontSize = tree.style.fontSize?.size ?? FontSize.medium.size;
 
-    tree.children.forEach((child) {
+    for (var child in tree.children) {
       if ((child.style.fontSize?.size ?? parentFontSize)! < 0) {
         child.style.fontSize =
             FontSize(parentFontSize! * -child.style.fontSize!.size!);
       }
 
       _processFontSize(child);
-    });
+    }
     return tree;
   }
 }
@@ -1031,6 +1125,7 @@ class RenderContext {
 /// A [ContainerSpan] can have a border, background color, height, width, padding, and margin
 /// and can represent either an INLINE or BLOCK-level element.
 class ContainerSpan extends StatelessWidget {
+  @override
   final AnchorKey? key;
   final Widget? child;
   final List<InlineSpan>? children;
@@ -1038,14 +1133,14 @@ class ContainerSpan extends StatelessWidget {
   final RenderContext newContext;
   final bool shrinkWrap;
 
-  ContainerSpan({
+  const ContainerSpan({
     this.key,
     this.child,
     this.children,
     required this.style,
     required this.newContext,
     this.shrinkWrap = false,
-  }): super(key: key);
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext _) {
@@ -1075,8 +1170,9 @@ class ContainerSpan extends StatelessWidget {
 class StyledText extends StatelessWidget {
   final InlineSpan textSpan;
   final Style style;
-  final double textScaleFactor;
+  final TextScaler? textScaler;
   final RenderContext renderContext;
+  @override
   final AnchorKey? key;
   final bool _selectable;
   final TextSelectionControls? selectionControls;
@@ -1085,23 +1181,23 @@ class StyledText extends StatelessWidget {
   const StyledText({
     required this.textSpan,
     required this.style,
-    this.textScaleFactor = 1.0,
+    this.textScaler,
     required this.renderContext,
     this.key,
     this.selectionControls,
     this.scrollPhysics,
-  }) : _selectable = false,
+  })  : _selectable = false,
         super(key: key);
 
   const StyledText.selectable({
     required TextSpan textSpan,
     required this.style,
-    this.textScaleFactor = 1.0,
+    this.textScaler,
     required this.renderContext,
     this.key,
     this.selectionControls,
     this.scrollPhysics,
-  }) : textSpan = textSpan,
+  })  : textSpan = textSpan,
         _selectable = true,
         super(key: key);
 
@@ -1113,7 +1209,7 @@ class StyledText extends StatelessWidget {
         style: style.generateTextStyle(),
         textAlign: style.textAlign,
         textDirection: style.direction,
-        textScaleFactor: textScaleFactor,
+        textScaler: textScaler,
         maxLines: style.maxLines,
         selectionControls: selectionControls,
         scrollPhysics: scrollPhysics,
@@ -1126,7 +1222,7 @@ class StyledText extends StatelessWidget {
         style: style.generateTextStyle(),
         textAlign: style.textAlign,
         textDirection: style.direction,
-        textScaleFactor: textScaleFactor,
+        textScaler: textScaler,
         maxLines: style.maxLines,
         overflow: style.textOverflow,
       ),
@@ -1134,7 +1230,8 @@ class StyledText extends StatelessWidget {
   }
 
   double? consumeExpandedBlock(Display? display, RenderContext context) {
-    if ((display == Display.BLOCK || display == Display.LIST_ITEM) && !renderContext.parser.shrinkWrap) {
+    if ((display == Display.BLOCK || display == Display.LIST_ITEM) &&
+        !renderContext.parser.shrinkWrap) {
       return double.infinity;
     }
     return null;
@@ -1143,9 +1240,10 @@ class StyledText extends StatelessWidget {
 
 extension IterateLetters on String {
   String nextLetter() {
-    String s = this.toLowerCase();
+    String s = toLowerCase();
     if (s == "z") {
-      return String.fromCharCode(s.codeUnitAt(0) - 25) + String.fromCharCode(s.codeUnitAt(0) - 25); // AA or aa
+      return String.fromCharCode(s.codeUnitAt(0) - 25) +
+          String.fromCharCode(s.codeUnitAt(0) - 25); // AA or aa
     } else {
       var lastChar = s.substring(s.length - 1);
       var sub = s.substring(0, s.length - 1);
@@ -1153,7 +1251,7 @@ extension IterateLetters on String {
         // If a string of length > 1 ends in Z/z,
         // increment the string (excluding the last Z/z) recursively,
         // and append A/a (depending on casing) to it
-        return sub.nextLetter() + 'a';
+        return '${sub.nextLetter()}a';
       } else {
         // (take till last char) append with (increment last char)
         return sub + String.fromCharCode(lastChar.codeUnitAt(0) + 1);
