@@ -1,13 +1,9 @@
-import 'dart:ui';
-
 import 'package:collection/collection.dart';
-import 'package:csslib/visitor.dart' as css;
 import 'package:csslib/parser.dart' as cssparser;
-import 'package:flutter/cupertino.dart';
+import 'package:csslib/visitor.dart' as css;
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:flutter_html/src/utils.dart';
-import 'package:flutter_html/style.dart';
 
 Style declarationsToStyle(Map<String, List<css.Expression>> declarations) {
   Style style = new Style();
@@ -15,53 +11,105 @@ Style declarationsToStyle(Map<String, List<css.Expression>> declarations) {
     if (value.isNotEmpty) {
       switch (property) {
         case 'background-color':
-          style.backgroundColor = ExpressionMapping.expressionToColor(value.first) ?? style.backgroundColor;
+          style.backgroundColor =
+              ExpressionMapping.expressionToColor(value.first) ??
+                  style.backgroundColor;
           break;
         case 'border':
-          List<css.LiteralTerm?>? borderWidths = value.whereType<css.LiteralTerm>().toList();
+          List<css.LiteralTerm?>? borderWidths =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.width], so make sure to remove those before passing it to [ExpressionMapping]
-          borderWidths.removeWhere((element) => element == null || (element.text != "thin"
-              && element.text != "medium" && element.text != "thick"
-              && !(element is css.LengthTerm) && !(element is css.PercentageTerm)
-              && !(element is css.EmTerm) && !(element is css.RemTerm)
-              && !(element is css.NumberTerm))
-          );
-          List<css.Expression?>? borderColors = value.where((element) => ExpressionMapping.expressionToColor(element) != null).toList();
-          List<css.LiteralTerm?>? potentialStyles = value.whereType<css.LiteralTerm>().toList();
+          borderWidths.removeWhere((element) =>
+              element == null ||
+              (element.text != "thin" &&
+                  element.text != "medium" &&
+                  element.text != "thick" &&
+                  !(element is css.LengthTerm) &&
+                  !(element is css.PercentageTerm) &&
+                  !(element is css.EmTerm) &&
+                  !(element is css.RemTerm) &&
+                  !(element is css.NumberTerm)));
+          List<css.Expression?>? borderColors = value
+              .where((element) =>
+                  ExpressionMapping.expressionToColor(element) != null)
+              .toList();
+          List<css.LiteralTerm?>? potentialStyles =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// Currently doesn't matter, as Flutter only supports "solid" or "none", but may support more in the future.
-          List<String> possibleBorderValues = ["dotted", "dashed", "solid", "double", "groove", "ridge", "inset", "outset", "none", "hidden"];
+          List<String> possibleBorderValues = [
+            "dotted",
+            "dashed",
+            "solid",
+            "double",
+            "groove",
+            "ridge",
+            "inset",
+            "outset",
+            "none",
+            "hidden"
+          ];
+
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.style], so make sure to remove those before passing it to [ExpressionMapping]
-          potentialStyles.removeWhere((element) => element == null || !possibleBorderValues.contains(element.text));
+          potentialStyles.removeWhere((element) =>
+              element == null || !possibleBorderValues.contains(element.text));
           List<css.LiteralTerm?>? borderStyles = potentialStyles;
-          style.border = ExpressionMapping.expressionToBorder(borderWidths, borderStyles, borderColors);
+          style.border = ExpressionMapping.expressionToBorder(
+              borderWidths, borderStyles, borderColors);
           break;
         case 'border-left':
-          List<css.LiteralTerm?>? borderWidths = value.whereType<css.LiteralTerm>().toList();
+          List<css.LiteralTerm?>? borderWidths =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.width], so make sure to remove those before passing it to [ExpressionMapping]
-          borderWidths.removeWhere((element) => element == null || (element.text != "thin"
-              && element.text != "medium" && element.text != "thick"
-              && !(element is css.LengthTerm) && !(element is css.PercentageTerm)
-              && !(element is css.EmTerm) && !(element is css.RemTerm)
-              && !(element is css.NumberTerm))
-          );
-          css.LiteralTerm? borderWidth = borderWidths.firstWhereOrNull((element) => element != null);
-          css.Expression? borderColor = value.firstWhereOrNull((element) => ExpressionMapping.expressionToColor(element) != null);
-          List<css.LiteralTerm?>? potentialStyles = value.whereType<css.LiteralTerm>().toList();
+          borderWidths.removeWhere((element) =>
+              element == null ||
+              (element.text != "thin" &&
+                  element.text != "medium" &&
+                  element.text != "thick" &&
+                  !(element is css.LengthTerm) &&
+                  !(element is css.PercentageTerm) &&
+                  !(element is css.EmTerm) &&
+                  !(element is css.RemTerm) &&
+                  !(element is css.NumberTerm)));
+          css.LiteralTerm? borderWidth =
+              borderWidths.firstWhereOrNull((element) => element != null);
+          css.Expression? borderColor = value.firstWhereOrNull((element) =>
+              ExpressionMapping.expressionToColor(element) != null);
+          List<css.LiteralTerm?>? potentialStyles =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// Currently doesn't matter, as Flutter only supports "solid" or "none", but may support more in the future.
-          List<String> possibleBorderValues = ["dotted", "dashed", "solid", "double", "groove", "ridge", "inset", "outset", "none", "hidden"];
+          List<String> possibleBorderValues = [
+            "dotted",
+            "dashed",
+            "solid",
+            "double",
+            "groove",
+            "ridge",
+            "inset",
+            "outset",
+            "none",
+            "hidden"
+          ];
+
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.style], so make sure to remove those before passing it to [ExpressionMapping]
-          potentialStyles.removeWhere((element) => element == null || !possibleBorderValues.contains(element.text));
+          potentialStyles.removeWhere((element) =>
+              element == null || !possibleBorderValues.contains(element.text));
           css.LiteralTerm borderStyle = potentialStyles.first!;
           Border newBorder = Border(
             left: style.border?.left.copyWith(
-              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-              color: ExpressionMapping.expressionToColor(borderColor),
-            ) ?? BorderSide(
-              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-              color: ExpressionMapping.expressionToColor(borderColor) ?? Colors.black,
-            ),
+                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+                  color: ExpressionMapping.expressionToColor(borderColor),
+                ) ??
+                BorderSide(
+                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+                  color: ExpressionMapping.expressionToColor(borderColor) ??
+                      Colors.black,
+                ),
             right: style.border?.right ?? BorderSide.none,
             top: style.border?.top ?? BorderSide.none,
             bottom: style.border?.bottom ?? BorderSide.none,
@@ -69,135 +117,226 @@ Style declarationsToStyle(Map<String, List<css.Expression>> declarations) {
           style.border = newBorder;
           break;
         case 'border-right':
-          List<css.LiteralTerm?>? borderWidths = value.whereType<css.LiteralTerm>().toList();
+          List<css.LiteralTerm?>? borderWidths =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.width], so make sure to remove those before passing it to [ExpressionMapping]
-          borderWidths.removeWhere((element) => element == null || (element.text != "thin"
-              && element.text != "medium" && element.text != "thick"
-              && !(element is css.LengthTerm) && !(element is css.PercentageTerm)
-              && !(element is css.EmTerm) && !(element is css.RemTerm)
-              && !(element is css.NumberTerm))
-          );
-          css.LiteralTerm? borderWidth = borderWidths.firstWhereOrNull((element) => element != null);
-          css.Expression? borderColor = value.firstWhereOrNull((element) => ExpressionMapping.expressionToColor(element) != null);
-          List<css.LiteralTerm?>? potentialStyles = value.whereType<css.LiteralTerm>().toList();
+          borderWidths.removeWhere((element) =>
+              element == null ||
+              (element.text != "thin" &&
+                  element.text != "medium" &&
+                  element.text != "thick" &&
+                  !(element is css.LengthTerm) &&
+                  !(element is css.PercentageTerm) &&
+                  !(element is css.EmTerm) &&
+                  !(element is css.RemTerm) &&
+                  !(element is css.NumberTerm)));
+          css.LiteralTerm? borderWidth =
+              borderWidths.firstWhereOrNull((element) => element != null);
+          css.Expression? borderColor = value.firstWhereOrNull((element) =>
+              ExpressionMapping.expressionToColor(element) != null);
+          List<css.LiteralTerm?>? potentialStyles =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// Currently doesn't matter, as Flutter only supports "solid" or "none", but may support more in the future.
-          List<String> possibleBorderValues = ["dotted", "dashed", "solid", "double", "groove", "ridge", "inset", "outset", "none", "hidden"];
+          List<String> possibleBorderValues = [
+            "dotted",
+            "dashed",
+            "solid",
+            "double",
+            "groove",
+            "ridge",
+            "inset",
+            "outset",
+            "none",
+            "hidden"
+          ];
+
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.style], so make sure to remove those before passing it to [ExpressionMapping]
-          potentialStyles.removeWhere((element) => element == null || !possibleBorderValues.contains(element.text));
+          potentialStyles.removeWhere((element) =>
+              element == null || !possibleBorderValues.contains(element.text));
           css.LiteralTerm borderStyle = potentialStyles.first!;
           Border newBorder = Border(
             left: style.border?.left ?? BorderSide.none,
             right: style.border?.right.copyWith(
-              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-              color: ExpressionMapping.expressionToColor(borderColor),
-            ) ?? BorderSide(
-              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-              color: ExpressionMapping.expressionToColor(borderColor) ?? Colors.black,
-            ),
+                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+                  color: ExpressionMapping.expressionToColor(borderColor),
+                ) ??
+                BorderSide(
+                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+                  color: ExpressionMapping.expressionToColor(borderColor) ??
+                      Colors.black,
+                ),
             top: style.border?.top ?? BorderSide.none,
             bottom: style.border?.bottom ?? BorderSide.none,
           );
           style.border = newBorder;
           break;
         case 'border-top':
-          List<css.LiteralTerm?>? borderWidths = value.whereType<css.LiteralTerm>().toList();
+          List<css.LiteralTerm?>? borderWidths =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.width], so make sure to remove those before passing it to [ExpressionMapping]
-          borderWidths.removeWhere((element) => element == null || (element.text != "thin"
-              && element.text != "medium" && element.text != "thick"
-              && !(element is css.LengthTerm) && !(element is css.PercentageTerm)
-              && !(element is css.EmTerm) && !(element is css.RemTerm)
-              && !(element is css.NumberTerm))
-          );
-          css.LiteralTerm? borderWidth = borderWidths.firstWhereOrNull((element) => element != null);
-          css.Expression? borderColor = value.firstWhereOrNull((element) => ExpressionMapping.expressionToColor(element) != null);
-          List<css.LiteralTerm?>? potentialStyles = value.whereType<css.LiteralTerm>().toList();
+          borderWidths.removeWhere((element) =>
+              element == null ||
+              (element.text != "thin" &&
+                  element.text != "medium" &&
+                  element.text != "thick" &&
+                  !(element is css.LengthTerm) &&
+                  !(element is css.PercentageTerm) &&
+                  !(element is css.EmTerm) &&
+                  !(element is css.RemTerm) &&
+                  !(element is css.NumberTerm)));
+          css.LiteralTerm? borderWidth =
+              borderWidths.firstWhereOrNull((element) => element != null);
+          css.Expression? borderColor = value.firstWhereOrNull((element) =>
+              ExpressionMapping.expressionToColor(element) != null);
+          List<css.LiteralTerm?>? potentialStyles =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// Currently doesn't matter, as Flutter only supports "solid" or "none", but may support more in the future.
-          List<String> possibleBorderValues = ["dotted", "dashed", "solid", "double", "groove", "ridge", "inset", "outset", "none", "hidden"];
+          List<String> possibleBorderValues = [
+            "dotted",
+            "dashed",
+            "solid",
+            "double",
+            "groove",
+            "ridge",
+            "inset",
+            "outset",
+            "none",
+            "hidden"
+          ];
+
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.style], so make sure to remove those before passing it to [ExpressionMapping]
-          potentialStyles.removeWhere((element) => element == null || !possibleBorderValues.contains(element.text));
+          potentialStyles.removeWhere((element) =>
+              element == null || !possibleBorderValues.contains(element.text));
           css.LiteralTerm borderStyle = potentialStyles.first!;
           Border newBorder = Border(
             left: style.border?.left ?? BorderSide.none,
             right: style.border?.right ?? BorderSide.none,
             top: style.border?.top.copyWith(
-              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-              color: ExpressionMapping.expressionToColor(borderColor),
-            ) ?? BorderSide(
-              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-              color: ExpressionMapping.expressionToColor(borderColor) ?? Colors.black,
-            ),
+                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+                  color: ExpressionMapping.expressionToColor(borderColor),
+                ) ??
+                BorderSide(
+                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+                  color: ExpressionMapping.expressionToColor(borderColor) ??
+                      Colors.black,
+                ),
             bottom: style.border?.bottom ?? BorderSide.none,
           );
           style.border = newBorder;
           break;
         case 'border-bottom':
-          List<css.LiteralTerm?>? borderWidths = value.whereType<css.LiteralTerm>().toList();
+          List<css.LiteralTerm?>? borderWidths =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.width], so make sure to remove those before passing it to [ExpressionMapping]
-          borderWidths.removeWhere((element) => element == null || (element.text != "thin"
-              && element.text != "medium" && element.text != "thick"
-              && !(element is css.LengthTerm) && !(element is css.PercentageTerm)
-              && !(element is css.EmTerm) && !(element is css.RemTerm)
-              && !(element is css.NumberTerm))
-          );
-          css.LiteralTerm? borderWidth = borderWidths.firstWhereOrNull((element) => element != null);
-          css.Expression? borderColor = value.firstWhereOrNull((element) => ExpressionMapping.expressionToColor(element) != null);
-          List<css.LiteralTerm?>? potentialStyles = value.whereType<css.LiteralTerm>().toList();
+          borderWidths.removeWhere((element) =>
+              element == null ||
+              (element.text != "thin" &&
+                  element.text != "medium" &&
+                  element.text != "thick" &&
+                  !(element is css.LengthTerm) &&
+                  !(element is css.PercentageTerm) &&
+                  !(element is css.EmTerm) &&
+                  !(element is css.RemTerm) &&
+                  !(element is css.NumberTerm)));
+          css.LiteralTerm? borderWidth =
+              borderWidths.firstWhereOrNull((element) => element != null);
+          css.Expression? borderColor = value.firstWhereOrNull((element) =>
+              ExpressionMapping.expressionToColor(element) != null);
+          List<css.LiteralTerm?>? potentialStyles =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// Currently doesn't matter, as Flutter only supports "solid" or "none", but may support more in the future.
-          List<String> possibleBorderValues = ["dotted", "dashed", "solid", "double", "groove", "ridge", "inset", "outset", "none", "hidden"];
+          List<String> possibleBorderValues = [
+            "dotted",
+            "dashed",
+            "solid",
+            "double",
+            "groove",
+            "ridge",
+            "inset",
+            "outset",
+            "none",
+            "hidden"
+          ];
+
           /// List<css.LiteralTerm> might include other values than the ones we want for [BorderSide.style], so make sure to remove those before passing it to [ExpressionMapping]
-          potentialStyles.removeWhere((element) => element == null || !possibleBorderValues.contains(element.text));
+          potentialStyles.removeWhere((element) =>
+              element == null || !possibleBorderValues.contains(element.text));
           css.LiteralTerm? borderStyle = potentialStyles.firstOrNull;
           Border newBorder = Border(
             left: style.border?.left ?? BorderSide.none,
             right: style.border?.right ?? BorderSide.none,
             top: style.border?.top ?? BorderSide.none,
             bottom: style.border?.bottom.copyWith(
-              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-              color: ExpressionMapping.expressionToColor(borderColor),
-            ) ?? BorderSide(
-              width: ExpressionMapping.expressionToBorderWidth(borderWidth),
-              style: ExpressionMapping.expressionToBorderStyle(borderStyle),
-              color: ExpressionMapping.expressionToColor(borderColor) ?? Colors.black,
-            ),
+                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+                  color: ExpressionMapping.expressionToColor(borderColor),
+                ) ??
+                BorderSide(
+                  width: ExpressionMapping.expressionToBorderWidth(borderWidth),
+                  style: ExpressionMapping.expressionToBorderStyle(borderStyle),
+                  color: ExpressionMapping.expressionToColor(borderColor) ??
+                      Colors.black,
+                ),
           );
           style.border = newBorder;
           break;
         case 'color':
-          style.color = ExpressionMapping.expressionToColor(value.first) ?? style.color;
+          style.color =
+              ExpressionMapping.expressionToColor(value.first) ?? style.color;
           break;
         case 'direction':
-          style.direction = ExpressionMapping.expressionToDirection(value.first);
+          style.direction =
+              ExpressionMapping.expressionToDirection(value.first);
           break;
         case 'display':
           style.display = ExpressionMapping.expressionToDisplay(value.first);
           break;
         case 'line-height':
-          style.lineHeight = ExpressionMapping.expressionToLineHeight(value.first);
+          style.lineHeight =
+              ExpressionMapping.expressionToLineHeight(value.first);
           break;
         case 'font-family':
-          style.fontFamily = ExpressionMapping.expressionToFontFamily(value.first) ?? style.fontFamily;
+          style.fontFamily =
+              ExpressionMapping.expressionToFontFamily(value.first) ??
+                  style.fontFamily;
           break;
         case 'font-feature-settings':
-          style.fontFeatureSettings = ExpressionMapping.expressionToFontFeatureSettings(value);
+          style.fontFeatureSettings =
+              ExpressionMapping.expressionToFontFeatureSettings(value);
           break;
         case 'font-size':
-          style.fontSize = ExpressionMapping.expressionToFontSize(value.first) ?? style.fontSize;
+          style.fontSize =
+              ExpressionMapping.expressionToFontSize(value.first) ??
+                  style.fontSize;
           break;
         case 'font-style':
-          style.fontStyle = ExpressionMapping.expressionToFontStyle(value.first);
+          style.fontStyle =
+              ExpressionMapping.expressionToFontStyle(value.first);
           break;
         case 'font-weight':
-          style.fontWeight = ExpressionMapping.expressionToFontWeight(value.first);
+          style.fontWeight =
+              ExpressionMapping.expressionToFontWeight(value.first);
           break;
         case 'list-style':
-          css.LiteralTerm? position = value.firstWhereOrNull((e) => e is css.LiteralTerm && (e.text == "outside" || e.text == "inside")) as css.LiteralTerm?;
-          css.UriTerm? image = value.firstWhereOrNull((e) => e is css.UriTerm) as css.UriTerm?;
-          css.LiteralTerm? type = value.firstWhereOrNull((e) => e is css.LiteralTerm && e.text != "outside" && e.text != "inside") as css.LiteralTerm?;
+          css.LiteralTerm? position = value.firstWhereOrNull((e) =>
+              e is css.LiteralTerm &&
+              (e.text == "outside" || e.text == "inside")) as css.LiteralTerm?;
+          css.UriTerm? image =
+              value.firstWhereOrNull((e) => e is css.UriTerm) as css.UriTerm?;
+          css.LiteralTerm? type = value.firstWhereOrNull((e) =>
+              e is css.LiteralTerm &&
+              e.text != "outside" &&
+              e.text != "inside") as css.LiteralTerm?;
           if (position != null) {
             switch (position.text) {
               case 'outside':
@@ -209,14 +348,20 @@ Style declarationsToStyle(Map<String, List<css.Expression>> declarations) {
             }
           }
           if (image != null) {
-            style.listStyleType = ExpressionMapping.expressionToListStyleType(image) ?? style.listStyleType;
+            style.listStyleType =
+                ExpressionMapping.expressionToListStyleType(image) ??
+                    style.listStyleType;
           } else if (type != null) {
-            style.listStyleType = ExpressionMapping.expressionToListStyleType(type) ?? style.listStyleType;
+            style.listStyleType =
+                ExpressionMapping.expressionToListStyleType(type) ??
+                    style.listStyleType;
           }
           break;
         case 'list-style-image':
           if (value.first is css.UriTerm) {
-            style.listStyleType = ExpressionMapping.expressionToListStyleType(value.first as css.UriTerm) ?? style.listStyleType;
+            style.listStyleType = ExpressionMapping.expressionToListStyleType(
+                    value.first as css.UriTerm) ??
+                style.listStyleType;
           }
           break;
         case 'list-style-position':
@@ -232,22 +377,29 @@ Style declarationsToStyle(Map<String, List<css.Expression>> declarations) {
           }
           break;
         case 'height':
-          style.height = ExpressionMapping.expressionToPaddingLength(value.first) ?? style.height;
+          style.height =
+              ExpressionMapping.expressionToPaddingLength(value.first) ??
+                  style.height;
           break;
         case 'list-style-type':
           if (value.first is css.LiteralTerm) {
-            style.listStyleType = ExpressionMapping.expressionToListStyleType(value.first as css.LiteralTerm) ?? style.listStyleType;
+            style.listStyleType = ExpressionMapping.expressionToListStyleType(
+                    value.first as css.LiteralTerm) ??
+                style.listStyleType;
           }
           break;
         case 'margin':
-          List<css.LiteralTerm>? marginLengths = value.whereType<css.LiteralTerm>().toList();
+          List<css.LiteralTerm>? marginLengths =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// List<css.LiteralTerm> might include other values than the ones we want for margin length, so make sure to remove those before passing it to [ExpressionMapping]
-          marginLengths.removeWhere((element) => !(element is css.LengthTerm)
-              && !(element is css.EmTerm)
-              && !(element is css.RemTerm)
-              && !(element is css.NumberTerm)
-          );
-          List<double?> margin = ExpressionMapping.expressionToPadding(marginLengths);
+          marginLengths.removeWhere((element) =>
+              !(element is css.LengthTerm) &&
+              !(element is css.EmTerm) &&
+              !(element is css.RemTerm) &&
+              !(element is css.NumberTerm));
+          List<double?> margin =
+              ExpressionMapping.expressionToPadding(marginLengths);
           style.margin = (style.margin ?? EdgeInsets.zero).copyWith(
             left: margin[0],
             right: margin[1],
@@ -272,14 +424,17 @@ Style declarationsToStyle(Map<String, List<css.Expression>> declarations) {
               bottom: ExpressionMapping.expressionToPaddingLength(value.first));
           break;
         case 'padding':
-          List<css.LiteralTerm>? paddingLengths = value.whereType<css.LiteralTerm>().toList();
+          List<css.LiteralTerm>? paddingLengths =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// List<css.LiteralTerm> might include other values than the ones we want for padding length, so make sure to remove those before passing it to [ExpressionMapping]
-          paddingLengths.removeWhere((element) => !(element is css.LengthTerm)
-              && !(element is css.EmTerm)
-              && !(element is css.RemTerm)
-              && !(element is css.NumberTerm)
-          );
-          List<double?> padding = ExpressionMapping.expressionToPadding(paddingLengths);
+          paddingLengths.removeWhere((element) =>
+              !(element is css.LengthTerm) &&
+              !(element is css.EmTerm) &&
+              !(element is css.RemTerm) &&
+              !(element is css.NumberTerm));
+          List<double?> padding =
+              ExpressionMapping.expressionToPadding(paddingLengths);
           style.padding = (style.padding ?? EdgeInsets.zero).copyWith(
             left: padding[0],
             right: padding[1],
@@ -304,36 +459,65 @@ Style declarationsToStyle(Map<String, List<css.Expression>> declarations) {
               bottom: ExpressionMapping.expressionToPaddingLength(value.first));
           break;
         case 'text-align':
-          style.textAlign = ExpressionMapping.expressionToTextAlign(value.first);
+          style.textAlign =
+              ExpressionMapping.expressionToTextAlign(value.first);
           break;
         case 'text-decoration':
-          List<css.LiteralTerm?>? textDecorationList = value.whereType<css.LiteralTerm>().toList();
+          List<css.LiteralTerm?>? textDecorationList =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// List<css.LiteralTerm> might include other values than the ones we want for [textDecorationList], so make sure to remove those before passing it to [ExpressionMapping]
-          textDecorationList.removeWhere((element) => element == null || (element.text != "none"
-              && element.text != "overline" && element.text != "underline" && element.text != "line-through"));
+          textDecorationList.removeWhere((element) =>
+              element == null ||
+              (element.text != "none" &&
+                  element.text != "overline" &&
+                  element.text != "underline" &&
+                  element.text != "line-through"));
           List<css.Expression?>? nullableList = value;
           css.Expression? textDecorationColor;
-          textDecorationColor = nullableList.firstWhereOrNull(
-                  (element) => element is css.HexColorTerm || element is css.FunctionTerm);
-          List<css.LiteralTerm?>? potentialStyles = value.whereType<css.LiteralTerm>().toList();
+          textDecorationColor = nullableList.firstWhereOrNull((element) =>
+              element is css.HexColorTerm || element is css.FunctionTerm);
+          List<css.LiteralTerm?>? potentialStyles =
+              value.whereType<css.LiteralTerm>().toList();
+
           /// List<css.LiteralTerm> might include other values than the ones we want for [textDecorationStyle], so make sure to remove those before passing it to [ExpressionMapping]
-          potentialStyles.removeWhere((element) => element == null || (element.text != "solid"
-              && element.text != "double" && element.text != "dashed" && element.text != "dotted" && element.text != "wavy"));
-          css.LiteralTerm? textDecorationStyle = potentialStyles.isNotEmpty ? potentialStyles.last : null;
-          style.textDecoration = ExpressionMapping.expressionToTextDecorationLine(textDecorationList);
-          if (textDecorationColor != null) style.textDecorationColor = ExpressionMapping.expressionToColor(textDecorationColor)
-              ?? style.textDecorationColor;
-          if (textDecorationStyle != null) style.textDecorationStyle = ExpressionMapping.expressionToTextDecorationStyle(textDecorationStyle);
+          potentialStyles.removeWhere((element) =>
+              element == null ||
+              (element.text != "solid" &&
+                  element.text != "double" &&
+                  element.text != "dashed" &&
+                  element.text != "dotted" &&
+                  element.text != "wavy"));
+          css.LiteralTerm? textDecorationStyle =
+              potentialStyles.isNotEmpty ? potentialStyles.last : null;
+          style.textDecoration =
+              ExpressionMapping.expressionToTextDecorationLine(
+                  textDecorationList);
+          if (textDecorationColor != null)
+            style.textDecorationColor =
+                ExpressionMapping.expressionToColor(textDecorationColor) ??
+                    style.textDecorationColor;
+          if (textDecorationStyle != null)
+            style.textDecorationStyle =
+                ExpressionMapping.expressionToTextDecorationStyle(
+                    textDecorationStyle);
           break;
         case 'text-decoration-color':
-          style.textDecorationColor = ExpressionMapping.expressionToColor(value.first) ?? style.textDecorationColor;
+          style.textDecorationColor =
+              ExpressionMapping.expressionToColor(value.first) ??
+                  style.textDecorationColor;
           break;
         case 'text-decoration-line':
-          List<css.LiteralTerm?>? textDecorationList = value.whereType<css.LiteralTerm>().toList();
-          style.textDecoration = ExpressionMapping.expressionToTextDecorationLine(textDecorationList);
+          List<css.LiteralTerm?>? textDecorationList =
+              value.whereType<css.LiteralTerm>().toList();
+          style.textDecoration =
+              ExpressionMapping.expressionToTextDecorationLine(
+                  textDecorationList);
           break;
         case 'text-decoration-style':
-          style.textDecorationStyle = ExpressionMapping.expressionToTextDecorationStyle(value.first as css.LiteralTerm);
+          style.textDecorationStyle =
+              ExpressionMapping.expressionToTextDecorationStyle(
+                  value.first as css.LiteralTerm);
           break;
         case 'text-shadow':
           style.textShadow = ExpressionMapping.expressionToTextShadow(value);
@@ -351,7 +535,9 @@ Style declarationsToStyle(Map<String, List<css.Expression>> declarations) {
           }
           break;
         case 'width':
-          style.width = ExpressionMapping.expressionToPaddingLength(value.first) ?? style.width;
+          style.width =
+              ExpressionMapping.expressionToPaddingLength(value.first) ??
+                  style.width;
           break;
       }
     }
@@ -374,7 +560,8 @@ Style? inlineCssToStyle(String? inlineStyle, OnCssParseError? errorHandler) {
   return null;
 }
 
-Map<String, Map<String, List<css.Expression>>> parseExternalCss(String css, OnCssParseError? errorHandler) {
+Map<String, Map<String, List<css.Expression>>> parseExternalCss(
+    String css, OnCssParseError? errorHandler) {
   var errors = <cssparser.Message>[];
   final sheet = cssparser.parse(css, errors: errors);
   if (errors.isEmpty) {
@@ -394,7 +581,8 @@ class DeclarationVisitor extends css.Visitor {
   late String _selector;
   late String _currentProperty;
 
-  Map<String, Map<String, List<css.Expression>>> getDeclarations(css.StyleSheet sheet) {
+  Map<String, Map<String, List<css.Expression>>> getDeclarations(
+      css.StyleSheet sheet) {
     sheet.topLevels.forEach((element) {
       if (element.span != null) {
         _selector = element.span!.text;
@@ -402,13 +590,15 @@ class DeclarationVisitor extends css.Visitor {
         if (_result[_selector] != null) {
           _properties.forEach((key, value) {
             if (_result[_selector]![key] != null) {
-              _result[_selector]![key]!.addAll(new List<css.Expression>.from(value));
+              _result[_selector]![key]!
+                  .addAll(new List<css.Expression>.from(value));
             } else {
               _result[_selector]![key] = new List<css.Expression>.from(value);
             }
           });
         } else {
-          _result[_selector] = new Map<String, List<css.Expression>>.from(_properties);
+          _result[_selector] =
+              new Map<String, List<css.Expression>>.from(_properties);
         }
         _properties.clear();
       }
@@ -435,8 +625,10 @@ class DeclarationVisitor extends css.Visitor {
 
 //Mapping functions
 class ExpressionMapping {
-
-  static Border expressionToBorder(List<css.Expression?>? borderWidths, List<css.LiteralTerm?>? borderStyles, List<css.Expression?>? borderColors) {
+  static Border expressionToBorder(
+      List<css.Expression?>? borderWidths,
+      List<css.LiteralTerm?>? borderStyles,
+      List<css.Expression?>? borderColors) {
     CustomBorderSide left = CustomBorderSide();
     CustomBorderSide top = CustomBorderSide();
     CustomBorderSide right = CustomBorderSide();
@@ -511,11 +703,22 @@ class ExpressionMapping {
       }
     }
     return Border(
-        top: BorderSide(width: top.width, color: top.color ?? Colors.black, style: top.style),
-        right: BorderSide(width: right.width, color: right.color ?? Colors.black, style: right.style),
-        bottom: BorderSide(width: bottom.width, color: bottom.color ?? Colors.black, style: bottom.style),
-        left: BorderSide(width: left.width, color: left.color ?? Colors.black, style: left.style)
-    );
+        top: BorderSide(
+            width: top.width,
+            color: top.color ?? Colors.black,
+            style: top.style),
+        right: BorderSide(
+            width: right.width,
+            color: right.color ?? Colors.black,
+            style: right.style),
+        bottom: BorderSide(
+            width: bottom.width,
+            color: bottom.color ?? Colors.black,
+            style: bottom.style),
+        left: BorderSide(
+            width: left.width,
+            color: left.color ?? Colors.black,
+            style: left.style));
   }
 
   static double expressionToBorderWidth(css.Expression? value) {
@@ -528,7 +731,9 @@ class ExpressionMapping {
     } else if (value is css.RemTerm) {
       return double.tryParse(value.text) ?? 1.0;
     } else if (value is css.LengthTerm) {
-      return double.tryParse(value.text.replaceAll(new RegExp(r'\s+(\d+\.\d+)\s+'), '')) ?? 1.0;
+      return double.tryParse(
+              value.text.replaceAll(new RegExp(r'\s+(\d+\.\d+)\s+'), '')) ??
+          1.0;
     } else if (value is css.LiteralTerm) {
       switch (value.text) {
         case "thin":
@@ -568,7 +773,7 @@ class ExpressionMapping {
 
   static TextDirection expressionToDirection(css.Expression value) {
     if (value is css.LiteralTerm) {
-      switch(value.text) {
+      switch (value.text) {
         case "ltr":
           return TextDirection.ltr;
         case "rtl":
@@ -580,7 +785,7 @@ class ExpressionMapping {
 
   static Display expressionToDisplay(css.Expression value) {
     if (value is css.LiteralTerm) {
-      switch(value.text) {
+      switch (value.text) {
         case 'block':
           return Display.BLOCK;
         case 'inline-block':
@@ -596,16 +801,25 @@ class ExpressionMapping {
     return Display.INLINE;
   }
 
-  static List<FontFeature> expressionToFontFeatureSettings(List<css.Expression> value) {
+  static List<FontFeature> expressionToFontFeatureSettings(
+      List<css.Expression> value) {
     List<FontFeature> fontFeatures = [];
     for (int i = 0; i < value.length; i++) {
       css.Expression exp = value[i];
       if (exp is css.LiteralTerm) {
-        if (exp.text != "on" && exp.text != "off" && exp.text != "1" && exp.text != "0") {
+        if (exp.text != "on" &&
+            exp.text != "off" &&
+            exp.text != "1" &&
+            exp.text != "0") {
           if (i < value.length - 1) {
-            css.Expression nextExp = value[i+1];
-            if (nextExp is css.LiteralTerm && (nextExp.text == "on" || nextExp.text == "off" || nextExp.text == "1" || nextExp.text == "0")) {
-              fontFeatures.add(FontFeature(exp.text, nextExp.text == "on" || nextExp.text == "1" ? 1 : 0));
+            css.Expression nextExp = value[i + 1];
+            if (nextExp is css.LiteralTerm &&
+                (nextExp.text == "on" ||
+                    nextExp.text == "off" ||
+                    nextExp.text == "1" ||
+                    nextExp.text == "0")) {
+              fontFeatures.add(FontFeature(exp.text,
+                  nextExp.text == "on" || nextExp.text == "1" ? 1 : 0));
             } else {
               fontFeatures.add(FontFeature.enable(exp.text));
             }
@@ -629,7 +843,8 @@ class ExpressionMapping {
     } else if (value is css.RemTerm) {
       return FontSize.rem(double.tryParse(value.text)!);
     } else if (value is css.LengthTerm) {
-      return FontSize(double.tryParse(value.text.replaceAll(new RegExp(r'\s+(\d+\.\d+)\s+'), '')));
+      return FontSize(double.tryParse(
+          value.text.replaceAll(new RegExp(r'\s+(\d+\.\d+)\s+'), '')));
     } else if (value is css.LiteralTerm) {
       switch (value.text) {
         case "xx-small":
@@ -653,7 +868,7 @@ class ExpressionMapping {
 
   static FontStyle expressionToFontStyle(css.Expression value) {
     if (value is css.LiteralTerm) {
-      switch(value.text) {
+      switch (value.text) {
         case "italic":
         case "oblique":
           return FontStyle.italic;
@@ -686,7 +901,7 @@ class ExpressionMapping {
           return FontWeight.w900;
       }
     } else if (value is css.LiteralTerm) {
-      switch(value.text) {
+      switch (value.text) {
         case "bold":
           return FontWeight.bold;
         case "bolder":
@@ -714,7 +929,10 @@ class ExpressionMapping {
     } else if (value is css.RemTerm) {
       return LineHeight.rem(double.tryParse(value.text)!);
     } else if (value is css.LengthTerm) {
-      return LineHeight(double.tryParse(value.text.replaceAll(new RegExp(r'\s+(\d+\.\d+)\s+'), '')), units: "length");
+      return LineHeight(
+          double.tryParse(
+              value.text.replaceAll(new RegExp(r'\s+(\d+\.\d+)\s+'), '')),
+          units: "length");
     }
     return LineHeight.normal;
   }
@@ -787,14 +1005,15 @@ class ExpressionMapping {
     } else if (value is css.RemTerm) {
       return double.tryParse(value.text);
     } else if (value is css.LengthTerm) {
-      return double.tryParse(value.text.replaceAll(new RegExp(r'\s+(\d+\.\d+)\s+'), ''));
+      return double.tryParse(
+          value.text.replaceAll(new RegExp(r'\s+(\d+\.\d+)\s+'), ''));
     }
     return null;
   }
 
   static TextAlign expressionToTextAlign(css.Expression value) {
     if (value is css.LiteralTerm) {
-      switch(value.text) {
+      switch (value.text) {
         case "center":
           return TextAlign.center;
         case "left":
@@ -812,11 +1031,12 @@ class ExpressionMapping {
     return TextAlign.start;
   }
 
-  static TextDecoration expressionToTextDecorationLine(List<css.LiteralTerm?> value) {
+  static TextDecoration expressionToTextDecorationLine(
+      List<css.LiteralTerm?> value) {
     List<TextDecoration> decorationList = [];
     for (css.LiteralTerm? term in value) {
       if (term != null) {
-        switch(term.text) {
+        switch (term.text) {
           case "overline":
             decorationList.add(TextDecoration.overline);
             break;
@@ -832,12 +1052,14 @@ class ExpressionMapping {
         }
       }
     }
-    if (decorationList.contains(TextDecoration.none)) decorationList = [TextDecoration.none];
+    if (decorationList.contains(TextDecoration.none))
+      decorationList = [TextDecoration.none];
     return TextDecoration.combine(decorationList);
   }
 
-  static TextDecorationStyle expressionToTextDecorationStyle(css.LiteralTerm value) {
-    switch(value.text) {
+  static TextDecorationStyle expressionToTextDecorationStyle(
+      css.LiteralTerm value) {
+    switch (value.text) {
       case "wavy":
         return TextDecorationStyle.wavy;
       case "dotted":
@@ -887,20 +1109,37 @@ class ExpressionMapping {
       });
       RegExp nonNumberRegex = RegExp(r'\s+(\d+\.\d+)\s+');
       if (offsetX is css.LiteralTerm && offsetY is css.LiteralTerm) {
-        if (color != null && ExpressionMapping.expressionToColor(color) != null) {
+        if (color != null &&
+            ExpressionMapping.expressionToColor(color) != null) {
           shadow.add(Shadow(
-              color: expressionToColor(color)!,
-              offset: Offset(
-                  double.tryParse((offsetX as css.LiteralTerm).text.replaceAll(nonNumberRegex, ''))!,
-                  double.tryParse((offsetY as css.LiteralTerm).text.replaceAll(nonNumberRegex, ''))!),
-              blurRadius: (blurRadius is css.LiteralTerm) ? double.tryParse((blurRadius as css.LiteralTerm).text.replaceAll(nonNumberRegex, ''))! : 0.0,
+            color: expressionToColor(color)!,
+            offset: Offset(
+                double.tryParse((offsetX as css.LiteralTerm)
+                    .text
+                    .replaceAll(nonNumberRegex, ''))!,
+                double.tryParse((offsetY as css.LiteralTerm)
+                    .text
+                    .replaceAll(nonNumberRegex, ''))!),
+            blurRadius: (blurRadius is css.LiteralTerm)
+                ? double.tryParse((blurRadius as css.LiteralTerm)
+                    .text
+                    .replaceAll(nonNumberRegex, ''))!
+                : 0.0,
           ));
         } else {
           shadow.add(Shadow(
-              offset: Offset(
-                  double.tryParse((offsetX as css.LiteralTerm).text.replaceAll(nonNumberRegex, ''))!,
-                  double.tryParse((offsetY as css.LiteralTerm).text.replaceAll(nonNumberRegex, ''))!),
-              blurRadius: (blurRadius is css.LiteralTerm) ? double.tryParse((blurRadius as css.LiteralTerm).text.replaceAll(nonNumberRegex, ''))! : 0.0,
+            offset: Offset(
+                double.tryParse((offsetX as css.LiteralTerm)
+                    .text
+                    .replaceAll(nonNumberRegex, ''))!,
+                double.tryParse((offsetY as css.LiteralTerm)
+                    .text
+                    .replaceAll(nonNumberRegex, ''))!),
+            blurRadius: (blurRadius is css.LiteralTerm)
+                ? double.tryParse((blurRadius as css.LiteralTerm)
+                    .text
+                    .replaceAll(nonNumberRegex, ''))!
+                : 0.0,
           ));
         }
       }
@@ -912,10 +1151,8 @@ class ExpressionMapping {
   static Color stringToColor(String _text) {
     var text = _text.replaceFirst('#', '');
     if (text.length == 3)
-      text = text.replaceAllMapped(
-          RegExp(r"[a-f]|\d", caseSensitive: false),
-          (match) => '${match.group(0)}${match.group(0)}'
-      );
+      text = text.replaceAllMapped(RegExp(r"[a-f]|\d", caseSensitive: false),
+          (match) => '${match.group(0)}${match.group(0)}');
     if (text.length > 6) {
       text = "0x" + text;
     } else {
@@ -928,7 +1165,7 @@ class ExpressionMapping {
     final rgbaText = text.replaceAll(')', '').replaceAll(' ', '');
     try {
       final rgbaValues =
-      rgbaText.split(',').map((value) => double.parse(value)).toList();
+          rgbaText.split(',').map((value) => double.parse(value)).toList();
       if (rgbaValues.length == 4) {
         return Color.fromRGBO(
           rgbaValues[0].toInt(),
@@ -955,10 +1192,13 @@ class ExpressionMapping {
     final hslValues = hslText.split(',').toList();
     List<double?> parsedHsl = [];
     hslValues.forEach((element) {
-      if (element.contains("%") && double.tryParse(element.replaceAll("%", "")) != null) {
+      if (element.contains("%") &&
+          double.tryParse(element.replaceAll("%", "")) != null) {
         parsedHsl.add(double.tryParse(element.replaceAll("%", ""))! * 0.01);
       } else {
-        if (element != hslValues.first && (double.tryParse(element) == null || double.tryParse(element)! > 1)) {
+        if (element != hslValues.first &&
+            (double.tryParse(element) == null ||
+                double.tryParse(element)! > 1)) {
           parsedHsl.add(null);
         } else {
           parsedHsl.add(double.tryParse(element));
@@ -966,16 +1206,24 @@ class ExpressionMapping {
       }
     });
     if (parsedHsl.length == 4 && !parsedHsl.contains(null)) {
-      return HSLColor.fromAHSL(parsedHsl.last!, parsedHsl.first!, parsedHsl[1]!, parsedHsl[2]!).toColor();
+      return HSLColor.fromAHSL(
+              parsedHsl.last!, parsedHsl.first!, parsedHsl[1]!, parsedHsl[2]!)
+          .toColor();
     } else if (parsedHsl.length == 3 && !parsedHsl.contains(null)) {
-      return HSLColor.fromAHSL(1.0, parsedHsl.first!, parsedHsl[1]!, parsedHsl.last!).toColor();
-    } else return Colors.black;
+      return HSLColor.fromAHSL(
+              1.0, parsedHsl.first!, parsedHsl[1]!, parsedHsl.last!)
+          .toColor();
+    } else
+      return Colors.black;
   }
 
   static Color? namedColorToColor(String text) {
-     String namedColor = namedColors.keys.firstWhere((element) => element.toLowerCase() == text.toLowerCase(), orElse: () => "");
-     if (namedColor != "") {
-       return stringToColor(namedColors[namedColor]!);
-     } else return null;
+    String namedColor = namedColors.keys.firstWhere(
+        (element) => element.toLowerCase() == text.toLowerCase(),
+        orElse: () => "");
+    if (namedColor != "") {
+      return stringToColor(namedColors[namedColor]!);
+    } else
+      return null;
   }
 }
