@@ -3,17 +3,15 @@ import 'dart:math';
 import 'package:chewie/chewie.dart';
 import 'package:chewie_audio/chewie_audio.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_html/html_parser.dart';
 import 'package:flutter_html/src/anchor.dart';
 import 'package:flutter_html/src/html_elements.dart';
 import 'package:flutter_html/src/navigation_delegate.dart';
 import 'package:flutter_html/src/utils.dart';
 import 'package:flutter_html/src/widgets/iframe_unsupported.dart'
-  if (dart.library.io) 'package:flutter_html/src/widgets/iframe_mobile.dart'
-  if (dart.library.html) 'package:flutter_html/src/widgets/iframe_web.dart';
+    if (dart.library.io) 'package:flutter_html/src/widgets/iframe_mobile.dart'
+    if (dart.library.html) 'package:flutter_html/src/widgets/iframe_web.dart';
 import 'package:flutter_html/style.dart';
-import 'package:flutter_math_fork/flutter_math.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:html/dom.dart' as dom;
 import 'package:video_player/video_player.dart';
@@ -32,7 +30,12 @@ abstract class ReplacedElement extends StyledElement {
     List<StyledElement>? children,
     dom.Element? node,
     this.alignment = PlaceholderAlignment.aboveBaseline,
-  }) : super(name: name, children: children ?? [], style: style, node: node, elementId: elementId);
+  }) : super(
+            name: name,
+            children: children ?? [],
+            style: style,
+            node: node,
+            elementId: elementId);
 
   static List<String?> parseMediaSources(List<dom.Element> elements) {
     return elements
@@ -55,7 +58,11 @@ class TextContentElement extends ReplacedElement {
     required this.text,
     this.node,
     dom.Element? element,
-  }) : super(name: "[text]", style: style, node: element, elementId: "[[No ID]]");
+  }) : super(
+            name: "[text]",
+            style: style,
+            node: element,
+            elementId: "[[No ID]]");
 
   @override
   String toString() {
@@ -77,27 +84,31 @@ class ImageContentElement extends ReplacedElement {
     required this.src,
     required this.alt,
     required dom.Element node,
-  }) : super(name: name, style: Style(), node: node, alignment: PlaceholderAlignment.middle, elementId: node.id);
+  }) : super(
+            name: name,
+            style: Style(),
+            node: node,
+            alignment: PlaceholderAlignment.middle,
+            elementId: node.id);
 
   @override
   Widget toWidget(RenderContext context) {
     for (final entry in context.parser.imageRenders.entries) {
       if (entry.key.call(attributes, element)) {
         final widget = entry.value.call(context, attributes, element);
-        return Builder(
-          builder: (buildContext) {
-            return GestureDetector(
-              key: AnchorKey.of(context.parser.key, this),
-              child: widget,
-              onTap: () {
-                if (MultipleTapGestureDetector.of(buildContext) != null) {
-                  MultipleTapGestureDetector.of(buildContext)!.onTap?.call();
-                }
-                context.parser.onImageTap?.call(src, context, attributes, element);
-              },
-            );
-          }
-        );
+        return Builder(builder: (buildContext) {
+          return GestureDetector(
+            key: AnchorKey.of(context.parser.key, this),
+            child: widget,
+            onTap: () {
+              if (MultipleTapGestureDetector.of(buildContext) != null) {
+                MultipleTapGestureDetector.of(buildContext)!.onTap?.call();
+              }
+              context.parser.onImageTap
+                  ?.call(src, context, attributes, element);
+            },
+          );
+        });
       }
     }
     return SizedBox(width: 0, height: 0);
@@ -128,11 +139,12 @@ class AudioContentElement extends ReplacedElement {
       key: AnchorKey.of(context.parser.key, this),
       width: context.style.width ?? 300,
       height: Theme.of(context.buildContext).platform == TargetPlatform.android
-          ? 48 : 75,
+          ? 48
+          : 75,
       child: ChewieAudio(
         controller: ChewieAudioController(
-          videoPlayerController: VideoPlayerController.network(
-            src.first ?? "",
+          videoPlayerController: VideoPlayerController.networkUrl(
+            Uri.parse(src.first ?? ""),
           ),
           autoPlay: autoplay,
           looping: loop,
@@ -178,8 +190,8 @@ class VideoContentElement extends ReplacedElement {
         key: AnchorKey.of(context.parser.key, this),
         child: Chewie(
           controller: ChewieController(
-            videoPlayerController: VideoPlayerController.network(
-              src.first ?? "",
+            videoPlayerController: VideoPlayerController.networkUrl(
+              Uri.parse(src.first ?? ""),
             ),
             placeholder: poster != null
                 ? Image.network(poster!)
@@ -208,7 +220,12 @@ class SvgContentElement extends ReplacedElement {
     required this.width,
     required this.height,
     required dom.Element node,
-  }) : super(name: name, style: Style(), node: node, elementId: node.id, alignment: PlaceholderAlignment.middle);
+  }) : super(
+            name: name,
+            style: Style(),
+            node: node,
+            elementId: node.id,
+            alignment: PlaceholderAlignment.middle);
 
   @override
   Widget toWidget(RenderContext context) {
@@ -222,7 +239,8 @@ class SvgContentElement extends ReplacedElement {
 }
 
 class EmptyContentElement extends ReplacedElement {
-  EmptyContentElement({String name = "empty"}) : super(name: name, style: Style(), elementId: "[[No ID]]");
+  EmptyContentElement({String name = "empty"})
+      : super(name: name, style: Style(), elementId: "[[No ID]]");
 
   @override
   Widget? toWidget(_) => null;
@@ -231,11 +249,16 @@ class EmptyContentElement extends ReplacedElement {
 class RubyElement extends ReplacedElement {
   dom.Element element;
 
-  RubyElement({
-    required this.element,
-    required List<StyledElement> children,
-    String name = "ruby"
-  }) : super(name: name, alignment: PlaceholderAlignment.middle, style: Style(), elementId: element.id, children: children);
+  RubyElement(
+      {required this.element,
+      required List<StyledElement> children,
+      String name = "ruby"})
+      : super(
+            name: name,
+            alignment: PlaceholderAlignment.middle,
+            style: Style(),
+            elementId: element.id,
+            children: children);
 
   @override
   Widget toWidget(RenderContext context) {
@@ -292,87 +315,6 @@ class RubyElement extends ReplacedElement {
   }
 }
 
-class MathElement extends ReplacedElement {
-  dom.Element element;
-  String? texStr;
-
-  MathElement({
-    required this.element,
-    this.texStr,
-    String name = "math",
-  }) : super(name: name, alignment: PlaceholderAlignment.middle, style: Style(display: Display.BLOCK), elementId: element.id);
-
-  @override
-  Widget toWidget(RenderContext context) {
-    texStr = parseMathRecursive(element, r'');
-    return Container(
-      width: context.parser.shrinkWrap ? null : MediaQuery.of(context.buildContext).size.width,
-      child: Math.tex(
-        texStr ?? '',
-        mathStyle: MathStyle.display,
-        textStyle: context.style.generateTextStyle(),
-        onErrorFallback: (FlutterMathException e) {
-          if (context.parser.onMathError != null) {
-            return context.parser.onMathError!.call(texStr ?? '', e.message, e.messageWithType);
-          } else {
-            return Text(e.message);
-          }
-        },
-      )
-    );
-  }
-
-  String parseMathRecursive(dom.Node node, String parsed) {
-    if (node is dom.Element) {
-      List<dom.Element> nodeList = node.nodes.whereType<dom.Element>().toList();
-      if (node.localName == "math" || node.localName == "mrow") {
-        nodeList.forEach((element) {
-          parsed = parseMathRecursive(element, parsed);
-        });
-      }
-      // note: munder, mover, and munderover do not support placing braces and other
-      // markings above/below elements, instead they are treated as super/subscripts for now.
-      if ((node.localName == "msup" || node.localName == "msub"
-          || node.localName == "munder" || node.localName == "mover") && nodeList.length == 2) {
-        parsed = parseMathRecursive(nodeList[0], parsed);
-        parsed = parseMathRecursive(nodeList[1],
-            parsed + "${node.localName == "msup" || node.localName == "mover" ? "^" : "_"}{") + "}";
-      }
-      if ((node.localName == "msubsup" || node.localName == "munderover") && nodeList.length == 3) {
-        parsed = parseMathRecursive(nodeList[0], parsed);
-        parsed = parseMathRecursive(nodeList[1], parsed + "_{") + "}";
-        parsed = parseMathRecursive(nodeList[2], parsed + "^{") + "}";
-      }
-      if (node.localName == "mfrac" && nodeList.length == 2) {
-        parsed = parseMathRecursive(nodeList[0], parsed + r"\frac{") + "}";
-        parsed = parseMathRecursive(nodeList[1], parsed + "{") + "}";
-      }
-      // note: doesn't support answer & intermediate steps
-      if (node.localName == "mlongdiv" && nodeList.length == 4) {
-        parsed = parseMathRecursive(nodeList[0], parsed);
-        parsed = parseMathRecursive(nodeList[2], parsed + r"\overline{)") + "}";
-      }
-      if (node.localName == "msqrt" && nodeList.length == 1) {
-        parsed = parseMathRecursive(nodeList[0], parsed + r"\sqrt{") + "}";
-      }
-      if (node.localName == "mroot" && nodeList.length == 2) {
-        parsed = parseMathRecursive(nodeList[1], parsed + r"\sqrt[") + "]";
-        parsed = parseMathRecursive(nodeList[0], parsed + "{") + "}";
-      }
-      if (node.localName == "mi" || node.localName == "mn" || node.localName == "mo") {
-        if (mathML2Tex.keys.contains(node.text.trim())) {
-          parsed = parsed + mathML2Tex[mathML2Tex.keys.firstWhere((e) => e == node.text.trim())]!;
-        } else if (node.text.startsWith("&") && node.text.endsWith(";")) {
-          parsed = parsed + node.text.trim().replaceFirst("&", r"\").substring(0, node.text.trim().length - 1);
-        } else {
-          parsed = parsed + node.text.trim();
-        }
-      }
-    }
-    return parsed;
-  }
-}
-
 ReplacedElement parseReplacedElement(
   dom.Element element,
   List<StyledElement> children,
@@ -398,19 +340,18 @@ ReplacedElement parseReplacedElement(
       );
     case "br":
       return TextContentElement(
-        text: "\n",
-        style: Style(whiteSpace: WhiteSpace.PRE),
-        element: element,
-        node: element
-      );
+          text: "\n",
+          style: Style(whiteSpace: WhiteSpace.PRE),
+          element: element,
+          node: element);
     case "iframe":
       return IframeContentElement(
-          name: "iframe",
-          src: element.attributes['src'],
-          width: double.tryParse(element.attributes['width'] ?? ""),
-          height: double.tryParse(element.attributes['height'] ?? ""),
-          navigationDelegate: navigationDelegateForIframe,
-          node: element,
+        name: "iframe",
+        src: element.attributes['src'],
+        width: double.tryParse(element.attributes['width'] ?? ""),
+        height: double.tryParse(element.attributes['height'] ?? ""),
+        navigationDelegate: navigationDelegateForIframe,
+        node: element,
       );
     case "img":
       return ImageContentElement(
@@ -452,11 +393,9 @@ ReplacedElement parseReplacedElement(
         element: element,
         children: children,
       );
-    case "math":
-      return MathElement(
-        element: element,
-      );
+
     default:
-      return EmptyContentElement(name: element.localName == null ? "[[No Name]]" : element.localName!);
+      return EmptyContentElement(
+          name: element.localName == null ? "[[No Name]]" : element.localName!);
   }
 }
